Move StringBuilderCache to shared
